refactor(livetalk): VOICEVOX クライアントを libs から services/livetalk/core へ移動
YAGNI レビューに基づく整理。現状 VOICEVOX クライアントを使うのはリブトークだけで、
他サービスの利用予定もないため、共通ライブラリ化は時期尚早と判断。複数サービスから
利用される必要が見えた段階で libs/voicevox-client/ へ昇格する。

- libs/voicevox-client/* を services/livetalk/core/src/voicevox/ へ移動
- @nagiyu/livetalk-core から VoicevoxClient / IVoiceClient / VOICEVOX_ERROR_MESSAGES を re-export
- web 側の dep / transpilePackages / jest moduleNameMapper / Dockerfile build step を整理
- CI（verify / deploy）から voicevox-client 個別ジョブを撤去
- docs/development/shared-libraries.md の libs ツリー記載・節を元に戻す
- core の tsconfig を lib: ES2022 に引き上げ（Error cause 利用のため）
- tasks/livetalk/design.md の配置記述を更新
