+++ .github/workflows/livetalk-deploy.yml
@@ -111,7 +111,7 @@ jobs:
         uses: ./.github/actions/build-web-app
         with:
           workspace: '@nagiyu/livetalk-web'
-          shared-workspaces: '@nagiyu/common,@nagiyu/browser,@nagiyu/ui,@nagiyu/nextjs,@nagiyu/voicevox-client,@nagiyu/livetalk-core'
+          shared-workspaces: '@nagiyu/common,@nagiyu/browser,@nagiyu/ui,@nagiyu/nextjs,@nagiyu/livetalk-core'
 
       - name: Configure AWS credentials
         uses: aws-actions/configure-aws-credentials@v4
