+++ .github/workflows/livetalk-verify.yml
@@ -36,9 +36,6 @@ jobs:
           node-version: '24'
           cache-dependency-path: './package-lock.json'
 
-      - name: Run ESLint on voicevox-client
-        run: npm run lint --workspace=@nagiyu/voicevox-client
-
       - name: Run ESLint on core
         run: npm run lint --workspace=@nagiyu/livetalk-core
 
@@ -58,9 +55,6 @@ jobs:
         with:
           node-version: '24'
           cache-dependency-path: './package-lock.json'
-
-      - name: Check voicevox-client formatting
-        run: npm run format:check --workspace=@nagiyu/voicevox-client
 
       - name: Check core formatting
         run: npm run format:check --workspace=@nagiyu/livetalk-core
@@ -127,7 +121,7 @@ jobs:
         uses: ./.github/actions/build-web-app
         with:
           workspace: '@nagiyu/livetalk-web'
-          shared-workspaces: '@nagiyu/common,@nagiyu/browser,@nagiyu/ui,@nagiyu/nextjs,@nagiyu/voicevox-client,@nagiyu/livetalk-core'
+          shared-workspaces: '@nagiyu/common,@nagiyu/browser,@nagiyu/ui,@nagiyu/nextjs,@nagiyu/livetalk-core'
 
   docker-build:
     name: Docker Build Verification
@@ -180,11 +174,7 @@ jobs:
           npm run build --workspace @nagiyu/browser
           npm run build --workspace @nagiyu/ui
           npm run build --workspace @nagiyu/nextjs
-          npm run build --workspace @nagiyu/voicevox-client
           npm run build --workspace @nagiyu/livetalk-core
-
-      - name: Run voicevox-client unit tests
-        run: npm run test --workspace @nagiyu/voicevox-client
 
       - name: Run core unit tests
         run: npm run test --workspace @nagiyu/livetalk-core
@@ -214,11 +204,7 @@ jobs:
           npm run build --workspace @nagiyu/browser
           npm run build --workspace @nagiyu/ui
           npm run build --workspace @nagiyu/nextjs
-          npm run build --workspace @nagiyu/voicevox-client
           npm run build --workspace @nagiyu/livetalk-core
-
-      - name: Run voicevox-client tests with coverage
-        run: npm run test:coverage --workspace @nagiyu/voicevox-client
 
       - name: Run core tests with coverage
         run: npm run test:coverage --workspace @nagiyu/livetalk-core
@@ -250,7 +236,6 @@ jobs:
           npm run build --workspace @nagiyu/browser
           npm run build --workspace @nagiyu/ui
           npm run build --workspace @nagiyu/nextjs
-          npm run build --workspace @nagiyu/voicevox-client
           npm run build --workspace @nagiyu/livetalk-core
 
       - name: Build web application
@@ -321,7 +306,6 @@ jobs:
           npm run build --workspace @nagiyu/browser
           npm run build --workspace @nagiyu/ui
           npm run build --workspace @nagiyu/nextjs
-          npm run build --workspace @nagiyu/voicevox-client
           npm run build --workspace @nagiyu/livetalk-core
 
       - name: Build web application
@@ -392,7 +376,6 @@ jobs:
           npm run build --workspace @nagiyu/browser
           npm run build --workspace @nagiyu/ui
           npm run build --workspace @nagiyu/nextjs
-          npm run build --workspace @nagiyu/voicevox-client
           npm run build --workspace @nagiyu/livetalk-core
 
       - name: Build web application
