--- libs/voicevox-client/eslint.config.mjs
@@ -1,3 +1,0 @@
-import baseConfig from '../../configs/eslint.config.base.mjs';
-
-export default [...baseConfig];
