--- libs/voicevox-client/src/index.ts
@@ -1,9 +1,0 @@
-/**
- * @nagiyu/voicevox-client
- *
- * VOICEVOX HTTP API クライアント。
- * リブトーク（LiveTalk）等のサービスから利用されることを想定。
- */
-
-export { VoicevoxClient, VOICEVOX_ERROR_MESSAGES } from './voicevox-client.js';
-export type { IVoiceClient, VoicevoxClientOptions } from './types.js';
