--- libs/voicevox-client/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../../configs/tsconfig.base.json",
-  "compilerOptions": {
-    "lib": ["ES2022"],
-    "composite": true,
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "outDir": "./dist",
-    "rootDir": ".",
-    "types": ["node"]
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist", "tests"],
-  "references": [{ "path": "../common" }]
-}
