--- services/livetalk/core/tests/unit/scaffold.test.ts
@@ -1,6 +1,0 @@
-describe('@nagiyu/livetalk-core scaffold', () => {
-  it('should be importable as an empty placeholder', async () => {
-    const mod = await import('../../src/index.js');
-    expect(mod).toBeDefined();
-  });
-});
