+++ services/livetalk/core/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../../configs/tsconfig.base.json",
   "compilerOptions": {
-    "lib": ["ES2020"],
+    "lib": ["ES2022"],
     "composite": true,
     "declaration": true,
     "outDir": "./dist",
