+++ services/livetalk/web/Dockerfile
@@ -27,7 +27,6 @@ RUN npm run build --workspace=@nagiyu/common
 RUN npm run build --workspace=@nagiyu/browser
 RUN npm run build --workspace=@nagiyu/ui
 RUN npm run build --workspace=@nagiyu/nextjs
-RUN npm run build --workspace=@nagiyu/voicevox-client
 RUN npm run build --workspace=@nagiyu/livetalk-core
 RUN npm run build --workspace=@nagiyu/livetalk-web
