+++ services/livetalk/web/jest.config.ts
@@ -17,7 +17,6 @@ const config: Config = {
     '^@nagiyu/common$': '<rootDir>/../../../libs/common/src/index.ts',
     '^@nagiyu/nextjs$': '<rootDir>/../../../libs/nextjs/src/index.ts',
     '^@nagiyu/livetalk-core$': '<rootDir>/../core/src/index.ts',
-    '^@nagiyu/voicevox-client$': '<rootDir>/../../../libs/voicevox-client/src/index.ts',
     '^(\\.{1,2}/.*)\\.js$': '$1',
   },
   testPathIgnorePatterns: ['/node_modules/', '/e2e/'],
