+++ services/livetalk/web/next.config.ts
@@ -11,7 +11,6 @@ const nextConfig: NextConfig = {
     '@nagiyu/common',
     '@nagiyu/nextjs',
     '@nagiyu/livetalk-core',
-    '@nagiyu/voicevox-client',
   ],
 };
